% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brapi_endpoints}
\alias{brapi_endpoints}
\title{List Available Endpoints}
\usage{
brapi_endpoints(con)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
}
\value{
A tibble with columns for endpoint service, method(s), and
version(s).
}
\description{
Queries the \verb{/serverinfo} endpoint to list which BrAPI calls the
server supports, along with their HTTP methods and versions.
}
\section{BrAPI endpoint}{

\code{GET /serverinfo} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Core/ServerInfo/ServerInfo_GET.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{contentType}, \code{dataType}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_endpoints(con)
}

}
