% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_germplasm_progeny}
\alias{brapi_germplasm_progeny}
\title{Get Germplasm Progeny}
\usage{
brapi_germplasm_progeny(con, germplasmDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{germplasmDbId}{Character. The unique germplasm identifier.}
}
\value{
A single-row tibble of the germplasm's pedigree node, with
\code{progeny} as a list-column of a tidy tibble of descendants. See
\code{\link[=brapi_pedigree]{brapi_pedigree()}}, which this function calls.
}
\description{
Get Germplasm Progeny
}
\section{BrAPI endpoint}{

\code{GET /pedigree} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Germplasm/Pedigree/Pedigree_GET_POST_PUT.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{accessionNumber}, \code{collection}, \code{familyCode}, \code{binomialName}, \code{genus},
\code{species}, \code{synonym}, \code{includeParents}, \code{includeSiblings},
\code{includeProgeny}, \code{includeFullTree}, \code{pedigreeDepth}, \code{progenyDepth}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
if (brapi_ping(con)) {
  brapi_germplasm_progeny(con, "germplasm1")
}
}

}
