% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_images}
\alias{brapi_images}
\title{List Images}
\usage{
brapi_images(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per image record.
}
\description{
List Images
}
\section{BrAPI endpoint}{

\code{GET /images} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Phenotyping/Images/Images_GET_POST.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{imageDbId}, \code{imageName}, \code{observationUnitDbId}, \code{observationDbId},
\code{descriptiveOntologyTerm}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_images(con)
}

}
