% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{brapi_login_oauth2}
\alias{brapi_login_oauth2}
\title{Login to a BrAPI Server with OAuth 2.0}
\usage{
brapi_login_oauth2(con, client_id, client_secret, authorize_url, access_url)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{client_id}{Character. OAuth client ID.}

\item{client_secret}{Character. OAuth client secret.}

\item{authorize_url}{Character. The authorization endpoint URL.}

\item{access_url}{Character. The token endpoint URL.}
}
\value{
A new \code{brapi_con} object with the token populated.
}
\description{
Performs an OAuth 2.0 authorization code flow or client credentials flow.
Returns an updated connection object with the Bearer token set.
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
if (brapi_ping(con)) {
  con <- brapi_login_oauth2(
    con,
    client_id = "brapi_client",
    client_secret = "brapi_secret",
    authorize_url = "https://test-server.brapi.org/brapi/v2/authorize",
    access_url = "https://test-server.brapi.org/brapi/v2/token"
  )
  con
}
}

}
