% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brapi_ping}
\alias{brapi_ping}
\title{Ping a BrAPI Server}
\usage{
brapi_ping(con)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
}
\value{
Logical. \code{TRUE} if the server responds, \code{FALSE} otherwise.
}
\description{
Tests whether the BrAPI server is reachable and responding.
}
\section{BrAPI endpoint}{

\code{GET /serverinfo} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Core/ServerInfo/ServerInfo_GET.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{contentType}, \code{dataType}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_ping(con)
}

}
