% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_program}
\alias{brapi_program}
\title{Get a Single Program by ID}
\usage{
brapi_program(con, programDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{programDbId}{Character. The unique program identifier.}
}
\value{
A single-row tibble with program details.
}
\description{
Get a Single Program by ID
}
\section{BrAPI endpoint}{

\code{GET /programs/{programDbId}} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Core/Programs/Programs_ProgramDbId_GET_PUT.yaml}{v2.1 specification}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
if (brapi_ping(con)) {
  brapi_program(con, "program1")
}
}

}
