% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_reference_sets}
\alias{brapi_reference_sets}
\title{List Reference Sets (Genome Assemblies)}
\usage{
brapi_reference_sets(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per reference set.
}
\description{
List Reference Sets (Genome Assemblies)
}
\section{BrAPI endpoint}{

\code{GET /referencesets} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Genotyping/ReferenceSets/ReferenceSets_GET.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{referenceSetDbId}, \code{accession}, \code{assemblyPUI}, \code{md5checksum}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_reference_sets(con)
}

}
