% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_samples}
\alias{brapi_samples}
\title{List Samples}
\usage{
brapi_samples(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters
(e.g. \code{studyDbId}, \code{germplasmDbId}, \code{observationUnitDbId}).}
}
\value{
A tibble with one row per sample.
}
\description{
List Samples
}
\section{BrAPI endpoint}{

\code{GET /samples} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Genotyping/Samples/Samples_GET_POST_PUT.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{sampleDbId}, \code{sampleName}, \code{sampleGroupDbId}, \code{observationUnitDbId},
\code{plateDbId}, \code{plateName}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
if (brapi_ping(con)) {
  brapi_samples(con)
}
}

}
