% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_search_calls}
\alias{brapi_search_calls}
\title{Search Genotype Calls}
\usage{
brapi_search_calls(con, variantSetDbIds = NULL, callSetDbIds = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{variantSetDbIds}{Character vector. Filter by variant set IDs.}

\item{callSetDbIds}{Character vector. Filter by call set IDs.}

\item{...}{Additional search body parameters.}
}
\value{
A tibble of matching genotype calls.
}
\description{
Search Genotype Calls
}
\section{BrAPI endpoint}{

\code{POST /search/calls} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Genotyping/Calls/Search_Calls_POST.yaml}{v2.1 specification}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
if (brapi_ping(con)) {
  brapi_search_calls(con, variantSetDbIds = "variantset1")
}
}

}
