% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_search_variables}
\alias{brapi_search_variables}
\title{Search Observation Variables}
\usage{
brapi_search_variables(con, traitClasses = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{traitClasses}{Character vector. Filter by trait class.}

\item{...}{Additional search body parameters.}
}
\value{
A tibble of matching observation variables.
}
\description{
Search Observation Variables
}
\section{BrAPI endpoint}{

\code{POST /search/variables} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Phenotyping/ObservationVariables/Search_Variables_POST.yaml}{v2.1 specification}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
if (brapi_ping(con)) {
  brapi_search_variables(con, traitClasses = "agronomic")
}
}

}
