% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_trial}
\alias{brapi_trial}
\title{Get a Single Trial by ID}
\usage{
brapi_trial(con, trialDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{trialDbId}{Character. The unique trial identifier.}
}
\value{
A single-row tibble with trial details.
}
\description{
Get a Single Trial by ID
}
\section{BrAPI endpoint}{

\code{GET /trials/{trialDbId}} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Core/Trials/Trials_TrialDbId_GET_PUT.yaml}{v2.1 specification}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_trial(con, "trial1")
}

}
