% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{is_brapi_con}
\alias{is_brapi_con}
\title{Test if an Object is a BrAPI Connection}
\usage{
is_brapi_con(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical.
}
\description{
Test if an Object is a BrAPI Connection
}
\keyword{internal}
