% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{parse_brapi_result}
\alias{parse_brapi_result}
\title{Internal: Parse BrAPI Result List into a Tibble}
\usage{
parse_brapi_result(data)
}
\arguments{
\item{data}{A list of named lists (one per record).}
}
\value{
A tibble.
}
\description{
Takes a list of result objects (each a named list) and flattens
them into a tibble. Nested lists are kept as list-columns.
}
\keyword{internal}
