% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.brapi_con}
\alias{print.brapi_con}
\title{Print a BrAPI Connection Object}
\usage{
\method{print}{brapi_con}(x, ...)
}
\arguments{
\item{x}{A \code{brapi_con} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a BrAPI Connection Object
}
\examples{
con <- brapi_connection("https://test-server.brapi.org")
print(con)

}
