% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_call_sets}
\alias{brapi_call_sets}
\title{List Call Sets (Samples with Genotype Data)}
\usage{
brapi_call_sets(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per call set.
}
\description{
List Call Sets (Samples with Genotype Data)
}
