% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_calls}
\alias{brapi_calls}
\title{List Genotype Calls}
\usage{
brapi_calls(con, variantSetDbId = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{variantSetDbId}{Character or NULL. Filter by variant set.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per genotype call.
}
\description{
List Genotype Calls
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_calls(con, variantSetDbId = "variantset1")
}

}
