% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_crosses}
\alias{brapi_crosses}
\title{List Crosses}
\usage{
brapi_crosses(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per cross.
}
\description{
List Crosses
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_crosses(con)
}

}
