% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_crossing_projects}
\alias{brapi_crossing_projects}
\title{List Crossing Projects}
\usage{
brapi_crossing_projects(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per crossing project.
}
\description{
List Crossing Projects
}
