% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brapi_endpoints}
\alias{brapi_endpoints}
\title{List Available Endpoints}
\usage{
brapi_endpoints(con)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
}
\value{
A tibble with columns for endpoint service, method(s), and version(s).
}
\description{
Queries the \verb{/serverinfo} endpoint to list which BrAPI calls the
server supports, along with their HTTP methods and versions.
}
\examples{
\dontrun{
con <- brapi_connection("https://test-server.brapi.org")
brapi_endpoints(con)
}

}
