% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_events}
\alias{brapi_events}
\title{List Events}
\usage{
brapi_events(con, studyDbId = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{studyDbId}{Character or NULL. Filter by study.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per event.
}
\description{
List Events
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_events(con, studyDbId = "study1")
}

}
