% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_germplasm}
\alias{brapi_germplasm}
\title{List Germplasm}
\usage{
brapi_germplasm(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters
(e.g. \code{commonCropName}, \code{germplasmName}, \code{studyDbId}).}
}
\value{
A tibble with one row per germplasm accession.
}
\description{
List Germplasm
}
