% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_germplasm_attributes}
\alias{brapi_germplasm_attributes}
\title{List Germplasm Attributes}
\usage{
brapi_germplasm_attributes(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per attribute definition.
}
\description{
List Germplasm Attributes
}
