% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_germplasm_detail}
\alias{brapi_germplasm_detail}
\title{Get a Single Germplasm by ID}
\usage{
brapi_germplasm_detail(con, germplasmDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{germplasmDbId}{Character. The unique germplasm identifier.}
}
\value{
A single-row tibble with germplasm details.
}
\description{
Get a Single Germplasm by ID
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_germplasm_detail(con, "germplasm1")
}

}
