% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_germplasm_pedigree}
\alias{brapi_germplasm_pedigree}
\title{Get Germplasm Pedigree}
\usage{
brapi_germplasm_pedigree(con, germplasmDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{germplasmDbId}{Character. The unique germplasm identifier.}
}
\value{
A tibble with pedigree information (parents, crosses).
}
\description{
Get Germplasm Pedigree
}
