% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_germplasm_progeny}
\alias{brapi_germplasm_progeny}
\title{Get Germplasm Progeny}
\usage{
brapi_germplasm_progeny(con, germplasmDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{germplasmDbId}{Character. The unique germplasm identifier.}
}
\value{
A single-row tibble of the germplasm's pedigree node, with
\code{progeny} as a list-column of a tidy tibble of descendants. See
\code{\link[=brapi_pedigree]{brapi_pedigree()}}, which this function calls.
}
\description{
Get Germplasm Progeny
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_germplasm_progeny(con, "germplasm1")
}

}
