% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_germplasm_progeny}
\alias{brapi_germplasm_progeny}
\title{Get Germplasm Progeny}
\usage{
brapi_germplasm_progeny(con, germplasmDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{germplasmDbId}{Character. The unique germplasm identifier.}
}
\value{
A tibble with progeny information.
}
\description{
Get Germplasm Progeny
}
