% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_lists}
\alias{brapi_lists}
\title{List Generic Lists}
\usage{
brapi_lists(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters (e.g. \code{listType}).}
}
\value{
A tibble with one row per list.
}
\description{
List Generic Lists
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_lists(con)
}

}
