% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_locations}
\alias{brapi_locations}
\title{List Locations}
\usage{
brapi_locations(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters (e.g. \code{locationType}).}
}
\value{
A tibble with one row per location.
}
\description{
List Locations
}
