% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{brapi_login}
\alias{brapi_login}
\title{Login to a BrAPI Server with Username and Password}
\usage{
brapi_login(con, username, password)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{username}{Character. Your username.}

\item{password}{Character. Your password.}
}
\value{
A new \code{brapi_con} object with the token populated.
}
\description{
Authenticates using the BrAPI \verb{/token} endpoint and returns an updated
connection object with the Bearer token set.
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
con <- brapi_login(con, "brapi_reader", "brapi_reader")
con
}

}
