% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{brapi_login_oauth2}
\alias{brapi_login_oauth2}
\title{Login to a BrAPI Server with OAuth 2.0}
\usage{
brapi_login_oauth2(con, client_id, client_secret, authorize_url, access_url)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{client_id}{Character. OAuth client ID.}

\item{client_secret}{Character. OAuth client secret.}

\item{authorize_url}{Character. The authorization endpoint URL.}

\item{access_url}{Character. The token endpoint URL.}
}
\value{
A new \code{brapi_con} object with the token populated.
}
\description{
Performs an OAuth 2.0 authorization code flow or client credentials flow.
Returns an updated connection object with the Bearer token set.
}
\examples{
\dontrun{
con <- brapi_connection("https://ebs.example.org")
con <- brapi_login_oauth2(
  con,
  client_id = "my_client_id",
  client_secret = "my_secret",
  authorize_url = "https://auth.example.org/oauth2/authorize",
  access_url = "https://auth.example.org/oauth2/token"
)
}

}
