% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_maps.R
\name{brapi_maps}
\alias{brapi_maps}
\title{List Genome Maps}
\usage{
brapi_maps(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per genome map.
}
\description{
List Genome Maps
}
\section{BrAPI endpoint}{

\code{GET /maps} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Genotyping/GenomeMaps/Maps_GET.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{mapDbId}, \code{mapPUI}, \code{scientificName}, \code{type}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
if (brapi_ping(con)) {
  brapi_maps(con)
}
}

}
