% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_methods}
\alias{brapi_methods}
\title{List Methods}
\usage{
brapi_methods(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per measurement method.
}
\description{
List Methods
}
