% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_observation_units}
\alias{brapi_observation_units}
\title{List Observation Units}
\usage{
brapi_observation_units(con, studyDbId = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{studyDbId}{Character or NULL. Filter by study.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per observation unit (plot/plant/sample).
}
\description{
List Observation Units
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_observation_units(con, studyDbId = "study1")
}

}
