% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_observation_variables}
\alias{brapi_observation_variables}
\title{List Observation Variables}
\usage{
brapi_observation_variables(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per variable definition.
}
\description{
Returns the ontology of observation variables (trait + method + scale).
}
