% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_observations}
\alias{brapi_observations}
\title{List Observations}
\usage{
brapi_observations(con, studyDbId = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{studyDbId}{Character or NULL. Filter by study.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per observation (trait measurement).
}
\description{
List Observations
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_observations(con, studyDbId = "study1")
}

}
