% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_people}
\alias{brapi_people}
\title{List People}
\usage{
brapi_people(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per person.
}
\description{
List People
}
