% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brapi_ping}
\alias{brapi_ping}
\title{Ping a BrAPI Server}
\usage{
brapi_ping(con)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
}
\value{
Logical. \code{TRUE} if the server responds, \code{FALSE} otherwise.
}
\description{
Tests whether the BrAPI server is reachable and responding.
}
\examples{
\dontrun{
con <- brapi_connection("https://test-server.brapi.org")
brapi_ping(con)
}

}
