% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_program}
\alias{brapi_program}
\title{Get a Single Program by ID}
\usage{
brapi_program(con, programDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{programDbId}{Character. The unique program identifier.}
}
\value{
A single-row tibble with program details.
}
\description{
Get a Single Program by ID
}
