% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_programs}
\alias{brapi_programs}
\title{List Breeding Programs}
\usage{
brapi_programs(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters passed to the API
(e.g. \code{commonCropName = "rice"}, \code{programName = "IRRI"}).}
}
\value{
A tibble with one row per program.
}
\description{
Retrieves a list of breeding programs from the BrAPI server.
}
\examples{
\dontrun{
con <- brapi_connection("https://test-server.brapi.org")
brapi_programs(con)
brapi_programs(con, commonCropName = "rice")
}

}
