% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_reference_sets}
\alias{brapi_reference_sets}
\title{List Reference Sets (Genome Assemblies)}
\usage{
brapi_reference_sets(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per reference set.
}
\description{
List Reference Sets (Genome Assemblies)
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_reference_sets(con)
}

}
