% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_references}
\alias{brapi_references}
\title{List References (Chromosomes/Contigs)}
\usage{
brapi_references(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per reference sequence.
}
\description{
List References (Chromosomes/Contigs)
}
\section{BrAPI endpoint}{

\code{GET /references} - see the
\href{https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Genotyping/References/References_GET.yaml}{v2.1 specification}.

Query parameters the specification defines, which may be passed
through \code{...}:

\code{referenceDbId}, \code{referenceSetDbId}, \code{accession}, \code{md5checksum},
\code{isDerived}, \code{minLength}, \code{maxLength}.
}

\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_references(con)
}

}
