% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_references}
\alias{brapi_references}
\title{List References (Chromosomes/Contigs)}
\usage{
brapi_references(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per reference sequence.
}
\description{
List References (Chromosomes/Contigs)
}
