% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_samples}
\alias{brapi_samples}
\title{List Samples}
\usage{
brapi_samples(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters
(e.g. \code{studyDbId}, \code{germplasmDbId}, \code{observationUnitDbId}).}
}
\value{
A tibble with one row per sample.
}
\description{
List Samples
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_samples(con)
}

}
