% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_scales}
\alias{brapi_scales}
\title{List Scales}
\usage{
brapi_scales(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per scale definition.
}
\description{
List Scales
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_scales(con)
}

}
