% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_search_germplasm}
\alias{brapi_search_germplasm}
\title{Search Germplasm}
\usage{
brapi_search_germplasm(
  con,
  germplasmNames = NULL,
  germplasmDbIds = NULL,
  commonCropNames = NULL,
  ...
)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{germplasmNames}{Character vector. Filter by germplasm names.}

\item{germplasmDbIds}{Character vector. Filter by database IDs.}

\item{commonCropNames}{Character vector. Filter by crop name.}

\item{...}{Additional body parameters for the search request.}
}
\value{
A tibble of matching germplasm records.
}
\description{
Performs a BrAPI search for germplasm records matching the given criteria.
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_search_germplasm(con, commonCropNames = "Tomatillo")
}

}
