% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_search_observations}
\alias{brapi_search_observations}
\title{Search Observations}
\usage{
brapi_search_observations(
  con,
  studyDbIds = NULL,
  observationVariableDbIds = NULL,
  ...
)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{studyDbIds}{Character vector. Filter by study IDs.}

\item{observationVariableDbIds}{Character vector. Filter by variable IDs.}

\item{...}{Additional search body parameters.}
}
\value{
A tibble of matching observations.
}
\description{
Search Observations
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_search_observations(con, studyDbIds = "study1")
}

}
