% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{brapi_search_variables}
\alias{brapi_search_variables}
\title{Search Observation Variables}
\usage{
brapi_search_variables(con, traitClasses = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{traitClasses}{Character vector. Filter by trait class.}

\item{...}{Additional search body parameters.}
}
\value{
A tibble of matching observation variables.
}
\description{
Search Observation Variables
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_search_variables(con, traitClasses = "agronomic")
}

}
