% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_search_variants}
\alias{brapi_search_variants}
\title{Search Variants}
\usage{
brapi_search_variants(con, variantSetDbIds = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{variantSetDbIds}{Character vector. Filter by variant set IDs.}

\item{...}{Additional search body parameters.}
}
\value{
A tibble of matching variants.
}
\description{
Search Variants
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_search_variants(con, variantSetDbIds = "variantset1")
}

}
