% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_seasons}
\alias{brapi_seasons}
\title{List Seasons}
\usage{
brapi_seasons(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters (e.g. \code{year}).}
}
\value{
A tibble with one row per season.
}
\description{
List Seasons
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_seasons(con)
}

}
