% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{brapi_seed_lots}
\alias{brapi_seed_lots}
\title{List Seed Lots}
\usage{
brapi_seed_lots(con, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per seed lot.
}
\description{
List Seed Lots
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_seed_lots(con)
}

}
