% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_server_info}
\alias{brapi_server_info}
\title{Get Server Info}
\usage{
brapi_server_info(con)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
}
\value{
A tibble of supported endpoints and their methods.
}
\description{
Returns a tibble of BrAPI calls supported by the server. Each row is one
supported endpoint with columns for service name, HTTP methods, BrAPI
versions, and content/data types.
}
\examples{
\dontrun{
con <- brapi_connection("https://test-server.brapi.org")
brapi_server_info(con)
}

}
