% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{brapi_set_token}
\alias{brapi_set_token}
\title{Manually Set an Authentication Token}
\usage{
brapi_set_token(con, token)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{token}{Character. The Bearer token string.}
}
\value{
A new \code{brapi_con} object with the token populated.
}
\description{
If you already have a token (e.g. from a web browser session), you can
set it directly without going through a login flow.
}
\examples{
con <- brapi_connection("https://test-server.brapi.org")
con <- brapi_set_token(con, "my_existing_token")

}
