% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_study}
\alias{brapi_study}
\title{Get a Single Study by ID}
\usage{
brapi_study(con, studyDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{studyDbId}{Character. The unique study identifier.}
}
\value{
A single-row tibble with study metadata.
}
\description{
Get a Single Study by ID
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_study(con, "study1")
}

}
