% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{brapi_trial}
\alias{brapi_trial}
\title{Get a Single Trial by ID}
\usage{
brapi_trial(con, trialDbId)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{trialDbId}{Character. The unique trial identifier.}
}
\value{
A single-row tibble with trial details.
}
\description{
Get a Single Trial by ID
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_trial(con, "trial1")
}

}
