% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_variant_sets}
\alias{brapi_variant_sets}
\title{List Variant Sets (Datasets)}
\usage{
brapi_variant_sets(con, studyDbId = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{studyDbId}{Character or NULL. Filter by study.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per variant set.
}
\description{
List Variant Sets (Datasets)
}
\examples{
\donttest{
con <- brapi_connection("https://test-server.brapi.org")
brapi_variant_sets(con)
}

}
