% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{brapi_variants}
\alias{brapi_variants}
\title{List Variants (Markers/SNPs)}
\usage{
brapi_variants(con, variantSetDbId = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}

\item{variantSetDbId}{Character or NULL. Filter by variant set.}

\item{...}{Additional query parameters.}
}
\value{
A tibble with one row per variant.
}
\description{
List Variants (Markers/SNPs)
}
