% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapiR2-package.R
\docType{package}
\name{brapiR2-package}
\alias{brapiR2}
\alias{brapiR2-package}
\title{brapiR2: A Tidyverse-Native Client for the 'BrAPI' v2 (Breeding API) Specification}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides pipe-friendly, stateless read access to the Breeding API ('BrAPI') v2.1 specification, an open community standard for plant breeding data interchange maintained by the BrAPI project \url{https://brapi.org}. Wraps 32 of the 37 'BrAPI' v2.1 entities across all four modules, Core, Germplasm, Phenotyping, and Genotyping, covering 49 of the specification's 138 retrieval ('GET' and search) endpoints and returning tidy tibbles ready for analysis. Write and update endpoints are out of scope by design. Features include automatic pagination, async search handling, response caching, parallel batch fetching, and convenience functions for genomic selection workflows (e.g. dosage matrix extraction). Designed for plant breeders and bioinformaticians who need programmatic access to plant breeding databases that implement the 'BrAPI' v2 specification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/josh45-source/brapiR2}
  \item \url{https://josh45-source.github.io/brapiR2/}
  \item Report bugs at \url{https://github.com/josh45-source/brapiR2/issues}
}

}
\author{
\strong{Maintainer}: Joash Joshua Ayo \email{joashjoshua789@gmail.com} (\href{https://orcid.org/0009-0007-1642-0172}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Joash Joshua Ayo \email{joashjoshua789@gmail.com} (\href{https://orcid.org/0009-0007-1642-0172}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item David Waring (David reviewed the package (v. 0.1.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/792>) [reviewer]
  \item Jenna Hershberger (Jenna reviewed the package (v. 0.1.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/792>) [reviewer]
}

}
\keyword{internal}
